AI_resume_screener/requirements.txt
google-generativeai
pypdf
python-dotenv
tiktoken
